+++ db/migrate/20190220171847_create_organizations.rb
@@ -3,7 +3,7 @@ class CreateOrganizations < ActiveRecord::Migration[5.2]
     create_table :organizations do |t|
       t.string :name
       t.string :description
-      t.integer :financial_need
+      t.bigint :financial_need
       t.timestamps
     end
   end
